requirements.txt
# Interview Sprint Platform - Dependencies
# Core CLI
rich>=13.7.0
click>=8.1.7
pyyaml>=6.0.1

# Testing
pytest>=8.0.0
pytest-timeout>=2.2.0

# Audio cues (cross-platform beep)
playsound>=1.3.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.29.0

# Progress tracking
tinydb>=4.8.0

# Date handling
python-dateutil>=2.8.2

